applying code review feedback
