+++ dyn/compat.py
@@ -49,6 +49,28 @@ if is_py2:
 
         return s
 
+    API_FMT = '%Y-%m-%dT%H:%M:%S'
+
+    def str_to_date(date_string):
+        """Convert a Message Manamgent API formatted string into a
+        standard python ``datetime.datetime`` object.  Note that this
+        object's tzinfo will not be set, in other words time zone is
+        ignored (this is due to a bug in python 2).
+        """
+        if date_string[-6:] == '+00:00':
+            date_string = date_string[:-6]
+        return datetime.strptime(date_string, API_FMT)
+
+    def date_to_str(date_obj):
+        """Convert a standard python ``datetime.datetime`` object to a
+        Dyn Message Management API formatted string.  Any supplied time
+        zone in the input will be ignored (due to a bug in python 2).
+        """
+        date_string = date_obj.strftime(API_FMT)
+        return date_string
+
+
+
 elif is_py3:
     from http.client import HTTPConnection, HTTPSConnection, HTTPException
     from urllib.parse import urlencode
@@ -64,3 +86,23 @@ elif is_py3:
 
     def force_unicode(s, encoding='UTF-8'):
         return str(s)
+
+    API_FMT = '%Y-%m-%dT%H:%M:%S%z'
+
+    def str_to_date(date_string):
+        """Convert a Message Manamgent API formatted string into a
+        standard python ``datetime.datetime`` object.
+        """
+        if date_string[-3] == ':':
+            date_string = date_string[:-3] + date_string[-2:]
+        return datetime.strptime(date_string, API_FMT)
+
+
+    def date_to_str(date_obj):
+        """Convert a standard python ``datetime.datetime`` object to a
+        Dyn Message Management API formatted string.
+        """
+        date_string = date_obj.strftime(API_FMT)
+        if date_string[-3] != ':':
+            date_string = date_string[:-2] + ':' + date_string[-2:]
+        return date_string
